Switch from android_glue to ndk-glue
This fixes compilation on Android. However, there is a caveat when
using it: ndk-glue doesn't have the necessary callbacks for glutin to
hook in to so that it can automatically handle destroying and creating
the context. The application must therefore ensure they only create a
Context following a winit Event::Resumed, and they destroy the context
when receiving an Event::Suspended.